requirements.txt
playwright
pandas
python-dotenv
beautifulsoup4
gspread
google-auth
supabase

requests
